requirements.txt
coverage==3.5.2
lettuce==0.1.35
nose==1.1.2
python-termstyle==0.1.10
sniffer==0.2.3
wsgiref==0.1.2
GitPython==0.3.2.RC1
